requirements.txt
beautifulsoup4==4.12.3
chardet==5.2.0
dj-database-url==3.0.1
Django==4.2.17
django-bootstrap-v5==1.0.11
django-taggit==6.1.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.1.0
python-decouple==3.8
reportlab==4.2.5
soupsieve==2.6
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.9.0
